+++ enume/example/lib/animal.dart
@@ -2,7 +2,7 @@ import 'package:enume/enume.dart';
 
 part 'animal.g.dart';
 
-@name
+@Enume()
 enum Animal {
   dog,
   cat,