+++ enume/example/lib/example.dart
@@ -1,17 +1,13 @@
 import 'package:example/animal.dart';
-import 'package:example/colors.dart';
-import 'package:example/difficulty.dart';
+import 'package:example/http_status.dart';
 
 void main() {
   // asString
   var animal = Animal.fish;
   print(animal.name);
 
-  // TypedEnums
-  print(Colors.blue.value);
+  var status = HttpStatus.ok;
+  print(status.value);
 
-  final experience = 45;
-  if (experience > Difficulty.hard.value) {
-    print('Lorem ipsum dolor sit amet...');
-  }
+  print(status.code);
 }