+++ example/lib/animal.dart
@@ -1,10 +1,11 @@
 import 'package:enume/enume.dart';
-import 'package:example/foo.dart';
 
 part 'animal.g.dart';
 
-@stringEnum
+@Enume()
 enum Animal {
-  @Value('value')
   dog,
+  cat,
+  fish,
+  fox,
 }