+++ example/lib/foo.dart
@@ -1,5 +1,0 @@
-class Foo {
-  final String s;
-
-  const Foo(this.s);
-}