+++ enume/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.1
+- Changed Annotation to @Enume()
+- Multiple @Value support 
+
 ## 1.0.1
 
 - README fix.